+++ vpt/lib/minecraft_util.py
@@ -5,7 +5,7 @@ from typing import Optional, Tuple
 import numpy as np
 import torch
 
-from vpt.action_head import (CategoricalActionHead, DiagGaussianActionHead,
+from vpt.lib.action_head import (CategoricalActionHead, DiagGaussianActionHead,
                              DictActionHead)
