+++ vpt/lib/misc.py
