+++ vpt/lib/normalize_ewma.py
