+++ vpt/lib/policy.py
@@ -8,14 +8,14 @@ from gym3.types import DictType
 from torch import nn
 from torch.nn import functional as F
 
-from vpt.action_head import make_action_head
-from vpt.action_mapping import CameraHierarchicalMapping
-from vpt.impala_cnn import ImpalaCNN
-from vpt.normalize_ewma import NormalizeEwma
-from vpt.scaled_mse_head import ScaledMSEHead
-from vpt.tree_util import tree_map
-from vpt.util import FanInInitReLULayer, ResidualRecurrentBlocks
-from vpt.misc import transpose
+from vpt.lib.action_head import make_action_head
+from vpt.lib.action_mapping import CameraHierarchicalMapping
+from vpt.lib.impala_cnn import ImpalaCNN
+from vpt.lib.normalize_ewma import NormalizeEwma
+from vpt.lib.scaled_mse_head import ScaledMSEHead
+from vpt.lib.tree_util import tree_map
+from vpt.lib.util import FanInInitReLULayer, ResidualRecurrentBlocks
+from vpt.lib.misc import transpose
 
 
 class ImgPreprocessing(nn.Module):
