+++ vpt/lib/torch_util.py
@@ -16,7 +16,7 @@ import torch.distributions as dis
 import torch.nn.functional as F
 from torch import nn
 
-import vpt.tree_util as tree_util
+import vpt.lib.tree_util as tree_util
 from vpt import misc
