+++ vpt/scripts/demo_rl_from_foundation_2x.sh
